allow set vae param dtype

+++ examples/opensora_hpcai/opensora/models/layers/blocks.py
@@ -58,7 +58,7 @@ class Attention(nn.Cell):
         # (b h n_q n_k)
         attn = ops.softmax(sim, axis=-1).astype(v.dtype)
         attn = self.attn_drop(attn)
-        out = ops.matmul(attn, v)
+        out = ops.matmul(attn.to(v.dtype), v)
 
         out = ops.reshape(out, (b, h, -1, d))
         # (b h n d) -> (b n h d)
